requirements.txt
backcall==0.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
monotonic==1.5
netifaces==0.10.4
numpy==1.21.1
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.9.3
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyX==0.15
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scapy==2.4.5
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
systemd-python==234
traitlets==4.3.3
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.1.8
xkit==0.0.0
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
psycopg2==2.9.3
python-dotenv==0.19.2
Werkzeug==2.0.2
